Move self signed certificate related code into its own file
